+++ plotoptix/_load_lib.py
@@ -7,9 +7,11 @@ Have a look at examples on GitHub: https://github.com/rnd-team-dev/plotoptix.
 
 import os, platform, sys
 
-from ctypes import cdll, CFUNCTYPE, POINTER, byref, cast, c_float, c_uint, c_int, c_long, c_longlong, c_bool, c_char_p, c_wchar_p, c_void_p
+from ctypes import cdll, CDLL, RTLD_GLOBAL, CFUNCTYPE, POINTER, byref, cast, c_float, c_uint, c_int, c_long, c_longlong, c_bool, c_char_p, c_wchar_p, c_void_p
 
 BIN_PATH = "bin"
+
+OIDN_VER = "2.2.1"
 
 PLATFORM = platform.system()
 if PLATFORM == "Linux":
@@ -535,6 +537,15 @@ class _ClrOptiX_v2:
     """
 
     def __init__(self):
+
+        try:
+            c_oidn_core = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, f"libOpenImageDenoise_core.so.{OIDN_VER}"), mode=RTLD_GLOBAL)
+            c_oidn_cuda = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, f"libOpenImageDenoise_device_cuda.so.{OIDN_VER}"), mode=RTLD_GLOBAL)
+            c_oidn = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, "libOpenImageDenoise.so"), mode=RTLD_GLOBAL)
+        except:
+            print(82 * "*"); print(82 * "*")
+            print("Open Image Denoiser library initialization failed.")
+            print(82 * "*"); print(82 * "*")
 
         try:
             c_encoder = cdll.LoadLibrary(os.path.join(os.path.dirname(__file__), BIN_PATH, "librndSharpEncoder.so"))
@@ -1185,6 +1196,15 @@ class _ClrOptiX_v3:
     """
 
     def __init__(self):
+
+        try:
+            c_oidn_core = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, f"libOpenImageDenoise_core.so.{OIDN_VER}"), mode=RTLD_GLOBAL)
+            c_oidn_cuda = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, f"libOpenImageDenoise_device_cuda.so.{OIDN_VER}"), mode=RTLD_GLOBAL)
+            c_oidn = CDLL(os.path.join(os.path.dirname(__file__), BIN_PATH, "libOpenImageDenoise.so"), mode=RTLD_GLOBAL)
+        except:
+            print(82 * "*"); print(82 * "*")
+            print("Open Image Denoiser library initialization failed.")
+            print(82 * "*"); print(82 * "*")
 
         try:
             c_encoder = cdll.LoadLibrary(os.path.join(os.path.dirname(__file__), BIN_PATH, "librndSharpEncoder.so"))