+++ plotoptix/install/__main__.py
@@ -8,10 +8,7 @@ import sys
 
 def main():
     if len(sys.argv) > 1:
-        if sys.argv[1] == "denoiser":
-            from plotoptix.install.denoiser import install_denoiser
-            result = install_denoiser()
-        elif sys.argv[1] == "examples":
+        if sys.argv[1] == "examples":
             from plotoptix.install.project import install_project
             result = install_project("examples.zip", "1Bdq7SnvI3fA12_-LoaF31h-d5E67T_32")
         elif sys.argv[1] == "moon":